+++ packages/terra-clinical-label-value-view/tests/wdio/label-value-view-spec.js
@@ -1,33 +1,25 @@
 Terra.describeViewports('Label Value View', ['tiny', 'small', 'medium', 'large', 'huge', 'enormous'], () => {
-  describe('when no value input is provided', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/default-label-value-view');
-    });
+  it('when no value input is provided', () => {
+    browser.url('/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/default-label-value-view');
 
-    Terra.it.validatesElement();
+    Terra.validates.element('no value');
   });
 
-  describe('when a text input is provided', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/text-value-label-value-view');
-    });
+  it('when a text input is provided', () => {
+    browser.url('/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/text-value-label-value-view');
 
-    Terra.it.validatesElement();
+    Terra.validates.element('text value');
   });
 
-  describe('when a node input is provided', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/node-value-label-value-view');
-    });
+  it('when a node input is provided', () => {
+    browser.url('/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/node-value-label-value-view');
 
-    Terra.it.validatesElement();
+    Terra.validates.element('node value');
   });
 
-  describe('when text and node inputs are provided', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/multiple-value-label-value-view');
-    });
+  it('when text and node inputs are provided', () => {
+    browser.url('/raw/tests/terra-clinical-label-value-view/clinical-label-value-view/multiple-value-label-value-view');
 
-    Terra.it.validatesElement();
+    Terra.validates.element('text and node value');
   });
 });