+++ packages/terra-clinical-label-value-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Update wdio tests to remove deprecated Terra.it helpers
+
 ## 3.22.0 - (August 25, 2020)
 
 * Changed